+++ cli/src/transactions.rs
@@ -42,17 +42,58 @@ pub fn payments(client: &Client, matches: &ArgMatches) -> Result<()> {
     let mut res = Ok(());
     pager.paginate(iter, |result| match result {
         Ok(op) => {
-            println!("ID:           {}", op.id());
-            if let &OperationKind::Payment(ref pymt) = op.kind() {
-                println!("To account:   {}", pymt.to());
-                println!("From account: {}", pymt.from());
-                println!("Asset");
-                println!("  Type:       {}", pymt.asset().asset_type());
-                println!("  Code:       {}", pymt.asset().code());
-                println!("  Issuer:     {}", pymt.asset().issuer());
-                println!("Amount:       {}", pymt.amount());
-            } else {
-                panic!("Did not receive Payment operation");
+            println!("ID:             {}", op.id());
+            match op.kind() {
+                &OperationKind::CreateAccount(ref create_account) => {
+                    println!("Operation Kind:   Create Account");
+                    println!("Account:          {}", create_account.account());
+                    println!("Funder:           {}", create_account.funder());
+                    println!("Starting Balance: {}", create_account.starting_balance());
+                }
+                &OperationKind::Payment(ref payment) => {
+                    println!("Operation Kind: Payment");
+                    println!("To account:     {}", payment.to());
+                    println!("From account:   {}", payment.from());
+                    println!("Asset Type:     {}", payment.asset().asset_type());
+                    println!("Asset Code:     {}", payment.asset().code());
+                    println!("Asset Issuer:   {}", payment.asset().issuer());
+                    println!("Amount:         {}", payment.amount());
+                }
+                &OperationKind::PathPayment(ref path_payment) => {
+                    println!("Operation Kind:           Path Payment");
+                    println!("To account:               {}", path_payment.to());
+                    println!("From account:             {}", path_payment.from());
+                    println!(
+                        "Source Asset Type:        {}",
+                        path_payment.source_asset().asset_type()
+                    );
+                    println!(
+                        "Source Asset Code:        {}",
+                        path_payment.source_asset().code()
+                    );
+                    println!(
+                        "Source Asset Issuer:      {}",
+                        path_payment.source_asset().issuer()
+                    );
+                    println!("Source Amount:            {}", path_payment.source_amount());
+                    println!(
+                        "Destination Asset Type:   {}",
+                        path_payment.destination_asset().asset_type()
+                    );
+                    println!(
+                        "Destination Asset Code:   {}",
+                        path_payment.destination_asset().code()
+                    );
+                    println!(
+                        "Destination Asset Issuer: {}",
+                        path_payment.destination_asset().issuer()
+                    );
+                    println!(
+                        "Destination Amount:       {}",
+                        path_payment.destination_amount()
+                    );
+                }
+                _ => println!("Something unexpected happened"),
             }
             println!();
         }