+++ cli/CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Transaction operations can be seen
 - Transaction effects can be seen
 - All trades can be seen
+- Ledger payments can be seen